requirements.txt
streamlit
PyPDF2
python-dotenv
pinecone-client
langchain-cohere
langchain
Cohere